Modularize negotiation stream handling
